+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Updated passwordless login to first try calling the API if it could guess a valid international phone number
 
+### Fixes
+
+-   Cleaned up some warnings shown in the console
+
 ## [0.27.0] - 2022-10-18
 
 ### Added